requirements.txt
click==6.7
awesome-slugify==1.6.5
requests==2.18.4
tinynetrc==1.1.0
pyyaml==3.12